refactor map helpers to own file, use observation end as start of slider
