+++ src/utils/helpers.ts
@@ -1,5 +1,4 @@
 import { Alert } from 'react-native';
-import moment from 'moment';
 import Geolocation from 'react-native-geolocation-service';
 import { TFunction } from 'react-i18next';
 
@@ -134,135 +133,4 @@ export const toStringWithDecimal = (
   if (Number.isNaN(input) || input === 0 || !input) return `0${separator}0`;
   if (Number.isInteger(input)) return `${input}${separator}0`;
   return input.toString().replace('.', separator);
-};
-
-type BoundingBox = {
-  CRS?: string;
-  minx: string;
-  miny: string;
-  maxx: string;
-  maxy: string;
-};
-
-type WmsLayer = {
-  Name: string;
-  Title: string;
-  Abstract: string;
-  CRS: string[];
-  BoundingBox: BoundingBox[];
-  Dimension: {
-    text: string;
-    name: string;
-    default: string;
-    units: string;
-  };
-};
-
-export const getRainRadarUrlsAndBounds = async (): Promise<
-  MapOverlay | undefined
-> => {
-  const toReturn = {
-    observation: {
-      bounds: undefined,
-      url: undefined,
-    },
-    forecast: {
-      bounds: undefined,
-      url: undefined,
-      start: undefined,
-    },
-  } as MapOverlay;
-
-  const sources = configJSON.map?.sources as { [key: string]: string };
-  const layers = configJSON.map?.layers?.find((layer) => layer.type === 'rain')
-    ?.sources as {
-    source: string;
-    layer: string;
-    type: string;
-  }[];
-
-  const rainRadarSources = Object.entries(sources)
-    .map(([key, value]) => layers.some((l) => l.source === key) && value)
-    .filter((x) => !!x);
-
-  await Promise.all(
-    rainRadarSources.map(async (url) => {
-      const capabilitiesUrl = `${url}&request=GetCapabilities`;
-      const textRes = await fetch(capabilitiesUrl).then((res) => res.text());
-
-      const obj = parse(textRes, {
-        attributeNamePrefix: '',
-        ignoreAttributes: false,
-        ignoreNameSpace: false,
-        textNodeName: 'text',
-      });
-
-      // layers array is nested in the response as Layer
-      const {
-        WMS_Capabilities: {
-          Capability: {
-            Layer: { Layer },
-          },
-        },
-      } = obj;
-
-      layers.forEach((layer) => {
-        const rainRadarLayer = Layer.find(
-          (l: WmsLayer) => l.Name === layer.layer
-        );
-
-        const { BoundingBox } = rainRadarLayer;
-        const [layerStart] = rainRadarLayer.Dimension.text.split('/');
-
-        const boundingBox84 = BoundingBox.find(
-          (box: BoundingBox) => box.CRS === 'CRS:84'
-        );
-        const { minx, miny, maxx, maxy } = boundingBox84;
-        const [numMinX, numMinY, numMaxX, numMaxY] = [
-          Number(minx),
-          Number(miny),
-          Number(maxx),
-          Number(maxy),
-        ];
-        const [minX, minY] = proj4('WGS84', 'EPSG:3857', [numMinX, numMinY]);
-        const [maxX, maxY] = proj4('WGS84', 'EPSG:3857', [numMaxX, numMaxY]);
-
-        const bbox = `${minX},${minY},${maxX},${maxY}`;
-
-        const overlayBounds = {
-          bottomLeft: [numMinY, numMinX],
-          bottomRight: [numMinY, numMaxX],
-          topLeft: [numMaxY, numMinX],
-          topRight: [numMaxY, numMaxX],
-        } as { [key: string]: [number, number] };
-
-        const query = new URLSearchParams({
-          service: 'WMS',
-          version: '1.3.0',
-          request: 'GetMap',
-          transparent: 'true',
-          layers: layer.layer,
-          bbox,
-          width: '727',
-          height: '1152',
-          format: 'image/png',
-          srs: 'EPSG:3857',
-        });
-
-        const overlayUrl = `${url}${query.toString()}`;
-
-        if (layer.type === 'observation') {
-          toReturn.observation.bounds = overlayBounds;
-          toReturn.observation.url = overlayUrl;
-        }
-        if (layer.type === 'forecast') {
-          toReturn.forecast.bounds = overlayBounds;
-          toReturn.forecast.url = overlayUrl;
-          toReturn.forecast.start = layerStart;
-        }
-      });
-    })
-  );
-
-  return toReturn;
 };
